Realized I poorly designed the edit use case, so split it up into two smaller ones
