+++ api/ExpressedRealms.Knowledges.UseCases.Tests.Unit/GetEditKnowledgeUseCaseTests.cs
@@ -1,4 +1,3 @@
-using ExpressedRealms.DB.Models.Knowledges;
 using ExpressedRealms.DB.Models.Knowledges.KnowledgeModels;
 using ExpressedRealms.Knowledges.Repository.Knowledges;
 using ExpressedRealms.Knowledges.UseCases.Knowledges.GetEditKnowledge;
@@ -15,7 +14,6 @@ public class GetEditKnowledgeUseCaseTests
     private readonly GetEditKnowledgeModel _model;
 
     private Knowledge KnowledgeDbModel { get; set; }
-    private List<KnowledgeType> KnowledgeTypesDbModel { get; set; }
 
     public GetEditKnowledgeUseCaseTests()
     {
@@ -28,27 +26,10 @@ public class GetEditKnowledgeUseCaseTests
             KnowledgeTypeId = 1,
         };
 
-        KnowledgeTypesDbModel = new List<KnowledgeType>()
-        {
-            new KnowledgeType()
-            {
-                Id = 1,
-                Name = "Test Knowledge Type 1",
-                Description = "Test Knowledge Type Description 1",
-            },
-            new KnowledgeType()
-            {
-                Id = 2,
-                Name = "Test Knowledge Type 2",
-                Description = "Test Knowledge Type Description 2",
-            },
-        };
-
         _repository = A.Fake<IKnowledgeRepository>();
 
         A.CallTo(() => _repository.IsExistingKnowledge(_model.Id)).Returns(true);
         A.CallTo(() => _repository.GetKnowledgeAsync(_model.Id)).Returns(KnowledgeDbModel);
-        A.CallTo(() => _repository.GetKnowledgeTypesAsync()).Returns(KnowledgeTypesDbModel);
 
         var validator = new GetEditKnowledgeModelValidator(_repository);
 
@@ -83,14 +64,6 @@ public class GetEditKnowledgeUseCaseTests
     }
 
     [Fact]
-    public async Task UseCase_Grabs_TheKnowledgeTypes()
-    {
-        await _useCase.ExecuteAsync(_model);
-
-        A.CallTo(() => _repository.GetKnowledgeTypesAsync()).MustHaveHappenedOnceExactly();
-    }
-
-    [Fact]
     public async Task UseCase_Returns_AllPropertiesForTheKnowledge()
     {
         var results = await _useCase.ExecuteAsync(_model);
@@ -99,22 +72,5 @@ public class GetEditKnowledgeUseCaseTests
         Assert.Equal(KnowledgeDbModel.Description, results.Value.Description);
         Assert.Equal(KnowledgeDbModel.KnowledgeTypeId, results.Value.KnowledgeTypeId);
         Assert.Equal(KnowledgeDbModel.Id, results.Value.Id);
-    }
-
-    [Fact]
-    public async Task UseCase_Returns_AvailableKnowledgeTypes()
-    {
-        var results = await _useCase.ExecuteAsync(_model);
-
-        var knowledgeTypes = KnowledgeTypesDbModel
-            .Select(x => new KnowledgeTypeModel()
-            {
-                Name = x.Name,
-                Description = x.Description,
-                Id = x.Id,
-            })
-            .ToList();
-
-        Assert.Equivalent(knowledgeTypes, results.Value.KnowledgeTypes);
     }
 }
