+++ api/ExpressedRealms.Knowledges.UseCases/KnowledgeTypes/GetKnowledgeTypes/GetKnowledgeTypeReturnModel.cs
@@ -1,0 +1,6 @@
+namespace ExpressedRealms.Knowledges.UseCases.KnowledgeTypes.GetKnowledgeTypes;
+
+public class GetKnowledgeTypeReturnModel
+{
+    public List<KnowledgeTypeModel> KnowledgeTypes { get; set; } = new();
+}
