+++ api/ExpressedRealms.Knowledges.UseCases/KnowledgeTypes/GetKnowledgeTypes/IGetKnowledgeTypesUseCase.cs
@@ -1,0 +1,9 @@
+using ExpressedRealms.Shared;
+using FluentResults;
+
+namespace ExpressedRealms.Knowledges.UseCases.KnowledgeTypes.GetKnowledgeTypes;
+
+public interface IGetKnowledgeTypesUseCase : IGenericUseCase<Result<GetKnowledgeTypeReturnModel>>
+{
+    
+}
