+++ api/ExpressedRealms.Knowledges.UseCases/Knowledges/GetEditKnowledge/GetEditKnowledgeReturnModel.cs
@@ -6,5 +6,4 @@ public class GetEditKnowledgeReturnModel
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
     public int KnowledgeTypeId { get; set; }
-    public List<KnowledgeTypeModel> KnowledgeTypes { get; set; } = new();
 }
